[Python] Split bindings in several .o files
