+++ bindings/python/algorithm/expose-algorithms.cpp
@@ -1,6 +1,5 @@
 //
 // Copyright (c) 2015-2016 CNRS
-// Copyright (c) 2015 Wandercraft, 86 rue de Paris 91400 Orsay, France.
 //
 // This file is part of Pinocchio
 // Pinocchio is free software: you can redistribute it
@@ -16,39 +15,31 @@
 // Pinocchio If not, see
 // <http://www.gnu.org/licenses/>.
 
-#ifndef __se3_python_python_hpp__
-#define __se3_python_python_hpp__
+#include "pinocchio/bindings/python/algorithm/algorithms.hpp"
 
 namespace se3
 {
   namespace python
   {
-    // Expose spatial classes
-    void exposeSE3();
-    void exposeForce();
-    void exposeMotion();
-    void exposeInertia();
-    void exposeExplog();
     
-    // Expose multibody classes
-    void exposeJoints();
-    void exposeModel();
-    void exposeFrame();
-    void exposeData();
-    
-    // Expose geometry module
+    void exposeAlgorithms()
+    {
+      exposeJointsAlgo();
+      exposeABA();
+      exposeCRBA();
+      exposeRNEA();
+      exposeCOM();
+      exposeFramesAlgo();
+      exposeEnergy();
+      exposeKinematics();
+      exposeDynamics();
+      exposeCAT();
+      exposeJacobian();
+      
 #ifdef WITH_HPP_FCL
-    void exposeGeometry();
+      exposeGeometryAlgo();
 #endif // ifdef WITH_HPP_FCL
+    }
     
-    // Expose parsers
-    void exposeParsers();
-    
-    // Expose algorithms
-    void exposeAlgorithms();
-
   } // namespace python
 } // namespace se3
-
-#endif // ifndef __se3_python_python_hpp__
-
